% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_first_stage_elasticity.R
\name{bootstrap_fs_perc_change}
\alias{bootstrap_fs_perc_change}
\title{Bootstrap SE Percentage Change Models}
\usage{
bootstrap_fs_perc_change(DT, form, y, x_main, x_int, B, quiet)
}
\arguments{
\item{DT}{a data.table}

\item{form}{fomula, first stage formula}

\item{y}{character vector, name of response variable(s) without months suffix}

\item{x_main}{character (default = "first_mo"), name of instrument}

\item{x_int}{character vector, names of variables to interact instrument with}

\item{B}{integer, number of bootstrap samples to use when calculating
standard errors}

\item{quiet}{logical (default = FALSE), if TRUE then does not print progress}
}
\value{
data.table with standard errors
}
\description{
Bootstrap SE Percentage Change Models
}
