% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enrollee_initial_features.R
\name{calc_initial_spending}
\alias{calc_initial_spending}
\title{Calculate Initial Spending (New Enrollees)}
\usage{
calc_initial_spending(DT, DT_id, id_vars, initial_days)
}
\arguments{
\item{DT}{data.table with all elements of id_vars as well as lab_prod}

\item{DT_id}{data.table data.table with all elements of id_vars}

\item{id_vars}{character vector, names of variables to aggregate on}

\item{initial_days}{integer, maximum number of days to use from enrollment}
}
\value{
data.table
}
\description{
Calculate Initial Spending (New Enrollees)
}
