% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_norm_perc.R
\name{calc_norm_perc}
\alias{calc_norm_perc}
\title{Calculate normalized percentage}
\usage{
calc_norm_perc(DT, x1, x2)
}
\arguments{
\item{DT}{data.table}

\item{x1}{character, the variable you want the normalized percentage of}

\item{x2}{character vector, the variables you want to normalize by}
}
\value{
data.table of normalized perctages
}
\description{
Calculate normalized percentage
}
