% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spending_by_month.R
\name{calc_spending_by_month}
\alias{calc_spending_by_month}
\title{Calculate New Enrollee Part D Spending by Month}
\usage{
calc_spending_by_month(DT, DT_id)
}
\arguments{
\item{DT}{data.table with the columns bene_id, year (relative to enrollment 
year, starting at 1), month, cost, and 
oop_cost}

\item{DT_id}{data.table with the column bene_id}
}
\value{
data.table with 7 columns: 
\item{bene_id}
\item{year}
\item{month}
\item{cost}{Total spending in month}
\item{cum_cost}{Total spending though month in year}
\item{oop_cost}{OOP spending in month}
\item{cum_oop_cost}{Total OOP spending through month in year}
}
\description{
calc_spend_by_month() is a function that calculates both the total/OOP
spending by calendar month, and cumulative amounts for each within a year.
}
