% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spending_by_month.R
\name{calc_spending_by_month_brand}
\alias{calc_spending_by_month_brand}
\title{Calculate New Enrollee Part D Spending by Month and Brand/Generic Type}
\usage{
calc_spending_by_month_brand(DT, DT_id)
}
\arguments{
\item{DT}{data.table with the columns bene_id, year (relative to enrollment 
year, starting at 1), month, cost, and oop_cost}

\item{DT_id}{data.table with the column bene_id}
}
\value{
data.table with spending/cumulative spending (total/oop) by year, 
 month, and brand type. 0 - generic, 1 - brand, 2 - unknown
}
\description{
Calculate New Enrollee Part D Spending by Month and Brand/Generic Type
}
