% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enrollee_initial_features.R
\name{calc_spending_in_class}
\alias{calc_spending_in_class}
\title{Calculate Spending within a Class}
\usage{
calc_spending_in_class(DT, cost_var, DT_id, id_vars, xwalk)
}
\arguments{
\item{DT}{data.table with all elements of id_vars as well as lab_prod and 
cost_var}

\item{cost_var}{character, name of the cost variable to use}

\item{DT_id}{data.table data.table with all elements of id_vars}

\item{id_vars}{character vector, names of variables to aggregate on}

\item{xwalk}{data.table with columns lab_prod as well as any indicators 
(usually atc1-4 or rxcui)}
}
\description{
Calculate Spending within a Class
}
