% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enrollee_initial_features.R
\name{calc_year_1_2_diff}
\alias{calc_year_1_2_diff}
\title{Calculate year 1/2 differences}
\usage{
calc_year_1_2_diff(DT, id_vars, time_var, year1_id, year2_id)
}
\arguments{
\item{DT}{data.table}

\item{id_vars}{character vector,}

\item{time_var}{character}

\item{year1_id}{character}

\item{year2_id}{character=}
}
\value{
data.table of year 1/2 differences
}
\description{
Calculate year 1/2 differences
}
