% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_risk.R
\name{calibrate_risk}
\alias{calibrate_risk}
\title{Calibrate Acute Event Prediction}
\usage{
calibrate_risk(DT, keep_vars, cont_risk_var, n_quant, event_var, treated_var)
}
\arguments{
\item{DT}{data.table}

\item{keep_vars}{character vector}

\item{cont_risk_var}{character}

\item{n_quant}{interger}

\item{event_var}{character}

\item{treated_var}{character}
}
\value{
data.table
}
\description{
Calibrate Acute Event Prediction
}
