% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfo.behavioral.R
\docType{package}
\name{cfo.behavioral}
\alias{cfo.behavioral}
\alias{cfo.behavioral-package}
\title{\code{} package}
\description{
R package for the work in progress "Behavioral Hazard and Patient Decision Making" by Chanddra, Flack, and Obermeyer.
}
