% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fit_dt.R
\name{clean_fit_dt}
\alias{clean_fit_dt}
\title{Format fit data.table}
\usage{
clean_fit_dt(
  dt,
  id_vars,
  est_var = "estimate",
  se_var = "std.error",
  p_var = "p.value",
  dig = 3
)
}
\arguments{
\item{dt}{data.table}

\item{id_vars}{character vector, name(s) of the idifiying featues of the 
estimate}

\item{est_var}{character, name of the estimate column}

\item{se_var}{character, name pf the standard error coulumn}

\item{p_var}{character, name of the  value column}

\item{dig}{integer, number of significant features to round estimate and 
standard error}
}
\value{
data.table ready to be printed with xtable
}
\description{
Format the estimates, standard errors, and significance stars for estimate
 tables to be printed in Latex.
}
