% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb.R
\name{comb}
\alias{comb}
\title{Combine list of data.table entries into one data.table}
\usage{
comb(x, ...)
}
\arguments{
\item{x}{list of data.tables}

\item{...}{additional arguments}
}
\value{
Combined data.table
}
\description{
Combine list of data.table entries into one data.table
}
