% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acute_event_outcomes.R
\name{create_diag_indicators}
\alias{create_diag_indicators}
\title{Create IP Diagnosis Indicators}
\usage{
create_diag_indicators(DT, DT_id, id_var, ami_codes, stroke_codes, diab_codes)
}
\arguments{
\item{DT}{data.table with ip claims}

\item{DT_id}{data.table with all id_vars}

\item{id_var}{character vecotr with names of coumns to aggregate on}

\item{ami_codes}{character vector of ami icd9 codes}

\item{stroke_codes}{character vector of stroke icd9 codes}

\item{diab_codes}{character vector of diabetes icd9 codes}
}
\value{
data.table with diagnoisis indicators
}
\description{
Create IP Diagnosis Indicators
}
