% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mortality_outcomes.R
\name{create_mortality_outcomes}
\alias{create_mortality_outcomes}
\title{Creat Mortality Outcomes}
\usage{
create_mortality_outcomes(DT_deaths, start_day = 0, int = 30, int_num = 12)
}
\arguments{
\item{DT_deaths}{data.table with bene_id, and from_days}

\item{start_day}{integer (default = 0), when to start counting outcomes from}

\item{int}{integer (default = 30), length of mortality intervals}

\item{int_num}{interger (default = 12), number of intervals to iterate 
through}
}
\value{
a data.table with bene_id, and indicators for cumulative mortality 
 and mortality hazard.
}
\description{
Creat Mortality Outcomes
}
