% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acute_event_outcomes.R
\name{create_prcdr_indicators}
\alias{create_prcdr_indicators}
\title{Create IP Procedure Indicators}
\usage{
create_prcdr_indicators(DT, DT_id, id_var)
}
\arguments{
\item{DT}{data.table with ip claims}

\item{DT_id}{data.table with all id_vars}

\item{id_var}{character vecotr with names of coumns to aggregate on}
}
\value{
data.table with procedure indicators
}
\description{
Create indicators for select inpatient procedures (tube and vent)
}
