% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_end_log_file.R
\name{end_log_file}
\alias{end_log_file}
\title{End a log file}
\usage{
end_log_file(log_file)
}
\arguments{
\item{log_file}{logicial, if TRUE then closes connection to current log file}
}
\description{
End a log file
}
