% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_first_stage_elasticity.R
\name{estimate_fs_perc_change}
\alias{estimate_fs_perc_change}
\title{Estimate FS Coefficient (Percentage Change)}
\usage{
estimate_fs_perc_change(DT, form, y, x_main, x_int)
}
\arguments{
\item{DT}{a data.table}

\item{form}{first stage formula}

\item{y}{character vector, name of response variable(s) without months suffix}

\item{x_main}{character (default = "first_mo"), name of instrument}

\item{x_int}{character vector, names of variables to interact instrument with}
}
\value{
List woth two elements
}
\description{
Estimate FS Coefficient (Percentage Change)
}
