% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_first_stage_elasticity.R
\name{estimate_fs_raw}
\alias{estimate_fs_raw}
\title{Estimate First Stage Coefficients}
\usage{
estimate_fs_raw(DT, y, month, x_main, x_int, form)
}
\arguments{
\item{DT}{a data.table}

\item{y}{character vector, name of response variable(s) without months suffix}

\item{month}{character, month of analysis (in form "year_month")}

\item{x_main}{character (default = "first_mo"), name of instrument}

\item{x_int}{character vector, names of variables to interact instrument with}

\item{form}{first stage formula}
}
\value{
data.table with first stage estimates and standard errors
}
\description{
Estimate First Stage Coefficients
}
