% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_1_2_diff.R
\name{estimate_year_1_2_diff}
\alias{estimate_year_1_2_diff}
\title{Estimate year 1 to 2 usage differencs}
\usage{
estimate_year_1_2_diff(DT, grid, year1_var, year2_var, x_main, x_int)
}
\arguments{
\item{DT}{data.table}

\item{grid}{data.table with combinations of jan/dec amounts}

\item{year1_var}{character, name of the column of the year 1 usage amount}

\item{year2_var}{character, name of the column of the year 2 usage amount}

\item{x_main}{character,}

\item{x_int}{character vector,}
}
\value{
data.table with first stage estimates for each combination in grid
}
\description{
Estimate the first stage of likelihood that people with a year 2 usage
 amount had a specific usage amount in year 1
}
