% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_1_2_diff.R
\name{estimate_year_1_2_diff_pooled}
\alias{estimate_year_1_2_diff_pooled}
\title{Estimate pooled year 1 to 2 usage differencs}
\usage{
estimate_year_1_2_diff_pooled(
  DT,
  diff_var,
  year2_var,
  u_diff_var,
  x_main,
  x_int,
  exclude = TRUE
)
}
\arguments{
\item{DT}{data.table}

\item{diff_var}{character, name of the column with the year 1 - year 2 amount}

\item{year2_var}{character, name of the column of the year 2 usage amount}

\item{u_diff_var}{numeric vecotr with all unique amounts of the year 1/2 
difference}

\item{x_main}{character,}

\item{x_int}{character vector,}

\item{exclude}{logical (default = TRUE), if TRUE, then excludes people with 
year 2 amoutns that are not compatible with the year 1 difference we are 
trying to estimate}
}
\value{
data.table with first stage estimates for each value in u_diff_var
}
\description{
Estimate the first stage of likelihood that people with a year 2 usage
 amount had a specific usage amount in year 1
}
