% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_zeros.R
\name{fill_in_zeros}
\alias{fill_in_zeros}
\title{Fill in NAs as 0}
\usage{
fill_in_zeros(DT, DT_id, id_name)
}
\arguments{
\item{DT}{a data.table with NA values}

\item{DT_id}{a data.table with all unique IDs}

\item{id_name}{name of id column(s)}
}
\value{
data.frame with added rows and 0s
}
\description{
Fill in NAs as 0
}
