% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_first_stage_elasticity.R
\name{fit_first_stage_price}
\alias{fit_first_stage_price}
\title{Fit first stage price}
\usage{
fit_first_stage_price(
  DT,
  keep_vars,
  cont_risk_var,
  n_quant_risk,
  atc_codes,
  rxcui_xwalk,
  price_by_drug,
  atc_ind,
  n_quant_price,
  x_main = "first_mo",
  x_int = c("pred_cut1", "hgih_risk_abs"),
  price_var,
  B = 100,
  return_data = FALSE,
  return_rxcui = FALSE
)
}
\arguments{
\item{DT}{data.table}

\item{keep_vars}{character vector}

\item{cont_risk_var}{character}

\item{n_quant_risk}{integer}

\item{atc_codes}{character vector}

\item{rxcui_xwalk}{data.table}

\item{price_by_drug}{data.table}

\item{atc_ind}{data.table}

\item{n_quant_price}{integer}

\item{x_main}{chatacer}

\item{x_int}{character vector}

\item{price_var}{character}

\item{B}{integer}

\item{return_data}{logical}

\item{return_rxcui}{logical}
}
\value{
data.table
}
\description{
Fit first stage price
}
