% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_to_dt.R
\name{fit_to_dt}
\alias{fit_to_dt}
\title{Format linear fit output as a data.table}
\usage{
fit_to_dt(fit, primary, interacts = NULL)
}
\arguments{
\item{fit}{the linear fit}

\item{primary}{the vaiable name of the rimary coefficient}

\item{interacts}{names of the interaction variables (factors)}
}
\value{
a data.table()
}
\description{
Format linear fit output as a data.table
}
