% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_train_ind.R
\name{get_train_ind}
\alias{get_train_ind}
\title{Get Indicies of Training Sample}
\usage{
get_train_ind(ids, population, pred_cat)
}
\arguments{
\item{ids}{data.table with keep variables}

\item{population}{character, name of population}

\item{pred_cat}{character, name of prediction category}
}
\value{
integer vector of the row indicies of observations in the training 
set
}
\description{
Get Indicies of Training Sample
}
