% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_2sls.R
\name{iterate_2sls}
\alias{iterate_2sls}
\title{Estimate 2SLS Mortality Model}
\usage{
iterate_2sls(DT, grid, max_cores)
}
\arguments{
\item{DT}{a data.table,}

\item{grid}{a data.table, estimation grid}

\item{max_cores}{integer, maximum number of parallels cores allowed to run at
once}
}
\value{
a data.table with the estimation options, 2SLS estimates, SEs, 
 p-values, CIs, etc...
}
\description{
Estimate 2SLS Mortality Model
}
