% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_first_stage_elasticity.R
\name{iterate_price}
\alias{iterate_price}
\title{Iterate Price}
\usage{
iterate_price(DT, form, price_change, B, price_var, x_main, x_int, quiet)
}
\arguments{
\item{DT}{a data.table}

\item{form}{formula, first stage formula}

\item{price_change}{data.table of price changes by drugs}

\item{B}{integer, number of bootstrap samples to use when calculating
standard errors}

\item{price_var}{character, name of price variable}

\item{x_main}{character (default = "first_mo"), name of instrument}

\item{x_int}{character vector, names of variables to interact instrument with}

\item{quiet}{logical (default = FALSE), if TRUE then does not print progress}
}
\value{
data.table of first stage estimates by price of drug
}
\description{
Iterate through different priced drugs in first stage percentage change 
 estimation
}
