% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_rf.R
\name{iterate_rf}
\alias{iterate_rf}
\title{Estimate reduced form models}
\usage{
iterate_rf(DT, grid, max_cores)
}
\arguments{
\item{DT}{a data.table with relevant model/sample option variables. Must have 
all the "keep" variables in grid, as well as the variables for all the 
options specified in grid. For example, if outcome = "mort" and 
outcome_period = 30, DT must have a column named "mort_30".}

\item{grid}{a data.table with estimation options (see below)}

\item{max_cores}{integer, maximum number of parallels cores allowed to run at
once}
}
\value{
data.table with the estimation options (from grid), RF estimates, 
 SEs, p-values, CIs, etc...
}
\description{
Iterate estimation of the mortality and utilization reduced forms over 
 multiple specificiations.
}
