% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_ri.R
\name{iterate_ri}
\alias{iterate_ri}
\title{Iterate Randomization Inference}
\usage{
iterate_ri(DT, grid, max_cores, controls)
}
\arguments{
\item{DT}{a data.table,}

\item{grid}{a data.table,}

\item{max_cores}{an integer,}

\item{controls, }{a character vector}
}
\value{
a data.table
}
\description{
iterate_ri() is a function that performs randomization inference on 
 instrumental variable models, and iterates over multiple sample
 specifications.
}
