% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit2prob.R
\name{logit2prob}
\alias{logit2prob}
\title{Convert logit prdictions to probabilities}
\usage{
logit2prob(logit)
}
\arguments{
\item{logit}{vector of logit preditions}
}
\value{
vector of predicted probabilities
}
\description{
Convert logit prdictions to probabilities
}
