% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enrollee_initial_features.R
\name{make_atc_indicators}
\alias{make_atc_indicators}
\title{Make ATC Indicators}
\usage{
make_atc_indicators(DT, DT_id, id_vars, atc_ind)
}
\arguments{
\item{DT}{data.table}

\item{DT_id}{data.table}

\item{id_vars}{character vector}

\item{atc_ind}{data.table}
}
\value{
data.table of atc indicators at the level of DT_id
}
\description{
Make ATC Indicators
}
