% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enrollee_initial_features.R
\name{make_class_indicators}
\alias{make_class_indicators}
\title{Make ATC Indicators}
\usage{
make_class_indicators(DT, DT_id, id_vars, xwalk)
}
\arguments{
\item{DT}{data.table with all elements of id_vars as well as lab_prod}

\item{DT_id}{data.table data.table with all elements of id_vars}

\item{id_vars}{character vector, names of variables to aggregate on}

\item{xwalk}{data.table with columns lab_prod as well as any indicators 
(usually atc1-4 or rxcui)}
}
\value{
data.table of atc indicators at the level of DT_id
}
\description{
Make ATC Indicators
}
