% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_first_stage_elasticity.R
\name{make_formula}
\alias{make_formula}
\title{Make Formula}
\usage{
make_formula(y, x_main, x_int)
}
\arguments{
\item{y}{character, name of the outcome variable}

\item{x_main}{character, name of the instrument}

\item{x_int}{character vector, name of the interaction variables}
}
\value{
formlula
}
\description{
Make Formula
}
