% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_first_stage_elasticity.R
\name{paste_factor}
\alias{paste_factor}
\title{Make a variable a factor for a formula}
\usage{
paste_factor(x)
}
\arguments{
\item{x}{string, name of variable in data.table}
}
\value{
string, "factor(x)"
}
\description{
Make a variable a factor for a formula
}
