% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polypharmacy.R
\docType{package}
\name{polypharmacy}
\alias{polypharmacy}
\alias{polypharmacy-package}
\title{\code{} package}
\description{
R package for the work in progress "Behavioral Hazard and Patient Decision Making" by Chanddra, Flack, and Obermeyer.
}
