% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_first_stage_elasticity.R
\name{prep_data}
\alias{prep_data}
\title{Prep Data}
\usage{
prep_data(DT, keep_vars, cont_risk_var = NULL, n_quant = 5)
}
\arguments{
\item{DT}{a data.table}

\item{keep_vars}{character vector, column names of binary keep variables}

\item{cont_risk_var}{character, name of continuous predicted risk variable}

\item{n_quant}{interger, numer of quantiles to cut the contrinuous risk 
variable in to. High risk is defined as the top quantile.}
}
\value{
subsetted data.table with new variables
}
\description{
Prep data for first stage estimation by subsettig based on binary "keep" 
 variables and defining "high_risk" if requested
}
