% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_first_stage_elasticity.R
\name{prep_fs_elasticity_data}
\alias{prep_fs_elasticity_data}
\title{Prep Data}
\usage{
prep_fs_elasticity_data(DT, keep_vars, cont_risk_var = NULL, n_quant = 5)
}
\arguments{
\item{DT}{a data.table}

\item{keep_vars}{character vector, column names for the binary keep variables}

\item{cont_risk_var}{character (default = NULL), name of the continuous
predicted risk variable}

\item{n_quant}{integer (default = 5), number of quntiles to break predicted
risk in to. High risk is defined as being in the top quantile.}
}
\value{
subsetted data.table with new variables
}
\description{
Prep data for first stage estimation by subsettig based on binary "keep"
 variables and defining "high_risk" if requested
}
