% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_first_stage_elasticity.R
\name{prep_price_data}
\alias{prep_price_data}
\title{Prep Price Data}
\usage{
prep_price_data(price_by_drug, rxcui_codes, n_quant, price_var)
}
\arguments{
\item{price_by_drug}{data.table with average price and number of observations
by g_rxcu code/coverage arm. Columns should be named as follows: g_rxcui, 
arm, oop_cost, obs}

\item{rxcui_codes}{character vector of g_rxcui code(s) of interest}

\item{n_quant}{number of quantiles to cut the price variable into}

\item{price_var}{price variable to use for cutting price into quantiles}
}
\value{
data.table
}
\description{
Prep Price Data
}
