% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_combine.R
\name{read_and_combine}
\alias{read_and_combine}
\title{Combine multiple years of data into one data.table}
\usage{
read_and_combine(lib_base_data, file, years, pct, loud = FALSE)
}
\arguments{
\item{lib_base_data}{the base directory for data}

\item{file}{base file name}

\item{years}{vecotor of years}

\item{pct}{string, sample percentage}

\item{loud, }{logical, if TRUE then prints years}
}
\value{
combined data.table
}
\description{
Combine multiple years of data into one data.table
}
