% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_month_level.R
\name{reshape_month_level}
\alias{reshape_month_level}
\title{Reseahpe wide data from BSF file to long}
\usage{
reshape_month_level(DT, id_vars, reshape_vars, value_name)
}
\arguments{
\item{DT}{data.table}

\item{id_vars}{string vector of id column name(s)}

\item{reshape_vars}{string vector, names of variables to reshape to long}

\item{value_name}{name for the new long variable}
}
\value{
long data.table
}
\description{
Reseahpe wide data from BSF file to long
}
