% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_first_stage_elasticity.R
\name{return_g_rxcui}
\alias{return_g_rxcui}
\title{Xwalk ATC4 to Generic RXCUI}
\usage{
return_g_rxcui(atc_codes, rxcui_xwalk, atc_ind)
}
\arguments{
\item{atc_codes}{character vector of atc code(s)}

\item{rxcui_xwalk}{data.table with 2 columns lab_prod (9 digit ndc) and 
g_rxcui (the generic rxcui code)}

\item{atc_ind}{data.table, xwalk between lab_prod and atc levels, atc levels
are indicators by column}
}
\value{
character vector of all generic rxcui codes associated with the
 given atc codes
}
\description{
Find all generic rxcui codes associated with a vector of atc codes
}
