% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsify.R
\name{sparsify}
\alias{sparsify}
\title{Iteratively make a matrix sparse}
\usage{
sparsify(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a sparse matrix
}
\description{
ITeratively make a matrix sparse to deal with memory issues as Matrix() 
 cannot make some matricies spare bacause theya re too big
}
