% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_sample.R
\name{subset_sample}
\alias{subset_sample}
\title{Subset to analystic sample}
\usage{
subset_sample(
  DT,
  subset_vars,
  progress = F,
  balance_vars = NULL,
  inst = "first_mo"
)
}
\arguments{
\item{DT}{a data.table}

\item{subset_vars}{variables to subset}

\item{progress}{logical, if TRUE prints where in subsetting process the
function is}

\item{balance_vars}{string character, if non-NULL, then estimates balance}

\item{inst}{chatacter, name of instrument to check balance on}
}
\value{
a list with three elements
\item{DT_subset}{Subsetted data.table}
\item{obs}{data.table with observations at each subsetting step}
\item{dt_bal}{data.table with first stage on balance variables at each 
              subset}
}
\description{
Subset to analystic sample
}
