% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{Bin a numeric variable}
\usage{
bin_variable(x, min = NULL, max = NULL, int = NULL, quant = NULL)
}
\arguments{
\item{x}{numeric vector to be binned}

\item{min}{where to start binning}

\item{max}{where to end binnig}

\item{int}{interval length of bins}

\item{quant}{number of quantiles}
}
\value{
vector of bins
}
\description{
bin_variable() is a function that creates bins of wither equal interval 
 length of equal size.
}
