% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cmean.R
\name{calc_cmean}
\alias{calc_cmean}
\title{Calculate conditional sample means and standard errors}
\usage{
calc_cmean(DT, y, x, se = F)
}
\arguments{
\item{DT}{a data.table}

\item{y}{a character vector of column names of data to calculate the group 
mean of}

\item{x}{a character vector of columns in data to group by}

\item{se}{logical, if TRUE output will calculate standard error and 
confidence intervals of mean alng with  number of observations of 
conditional sample mean}
}
\value{
The data.frame/data.table of conditional sample means
}
\description{
calc_cmean() is a function that calculates sample means (plus
 standard errors and confidence intervals) by group.
}
\examples{
library(data.table)
data <- data.table(x1 = sample(1:5, 1000, replace = TRUE),
                   x2 = sample(1:2, 1000, replace = TRUE))
data[, y := 2*x1 + 4*x2 + rnorm(mean = 0, sd = 2, n = 1000)]

calc_cmean(data, y = c("y"), x = c("x1", "x2"))

}
