% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enrollee_initial_features.R
\name{calc_initial_spending}
\alias{calc_initial_spending}
\title{Calculate Initial Spending (New Enrollees)}
\usage{
calc_initial_spending(DT, DT_id, id_vars, initial_days)
}
\arguments{
\item{DT}{data.table}

\item{DT_id}{data.table}

\item{id_vars}{character vector}

\item{initial_days}{integer vector}
}
\value{
data.table
}
\description{
Calculate Initial Spending (New Enrollees)
}
