% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enrollee_initial_features.R
\name{calc_year_1_2_diff}
\alias{calc_year_1_2_diff}
\title{Calculate year 1/2 differences}
\usage{
calc_year_1_2_diff(DT, id_vars, time_var, year1_id, year2_id)
}
\arguments{
\item{DT}{data.table with id_vars, time_var, and year 1/2 utilization 
variables}

\item{id_vars}{character vector, names of id variables}

\item{time_var}{character, name of time variable}

\item{year1_id}{character, value of time_var for year 1 observations}

\item{year2_id}{character, value of time_var for year 2 observations}
}
\value{
data.table of year 1/2 differences
}
\description{
Calculate year 1/2 differences
}
