% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_risk.R
\name{calibrate_risk}
\alias{calibrate_risk}
\title{Calibrate Acute Event Prediction}
\usage{
calibrate_risk(DT, keep_vars, cont_risk_var, n_quant, event_var, treated_var)
}
\arguments{
\item{DT}{data.table with keep_vars, cont_risk_var, event_var, and 
treated_var}

\item{keep_vars}{character vector, names of subsetting variables (e.g. 
"keep_jan")}

\item{cont_risk_var}{character, name of contrinuous risk prediction variable
(e.g. "comp_card_untreated_statin")}

\item{n_quant}{interger, number of quantiles to break cont_risk_var in to}

\item{event_var}{character, name of the realized event variable (e.g. "ami")}

\item{treated_var}{character, name of the binary treatment indicator 
variable}
}
\value{
data.table with realized event rate by treatment status and predicted
 risk quantile
}
\description{
Calibrate Acute Event Prediction
}
