% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_to_dt.R
\name{fit_to_dt}
\alias{fit_to_dt}
\title{Format linear fit output as a data.table}
\usage{
fit_to_dt(fit, primary, interacts = NULL)
}
\arguments{
\item{fit}{the linear fit (usually lm, lm_robust, or iv_robust)}

\item{primary}{the vaiable name of the rimary coefficient}

\item{interacts}{names of the interaction variables (factors)}
}
\value{
data.table with all coefficients of primary variable
}
\description{
fit_to_dt() is a function that converts the results of a model into a 
 data.table with easily manipulable columns of interaction labels.
}
\examples{
DT <- as.data.table(mtcars)
fit <- lm(mpg ~ wt:factor(cyl) + factor(cyl), data = DT)
dt_fit <- fit_to_dt(fit, primary = "wt", interacts = "cyl")

}
