% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enrollee_initial_features.R
\name{indicate_initial_atc}
\alias{indicate_initial_atc}
\title{Create Initial ATC Indicators (New Enrollees)}
\usage{
indicate_initial_atc(DT, DT_id, atc_ind, initial_days)
}
\arguments{
\item{DT}{data.table}

\item{DT_id}{data.table}

\item{atc_ind}{data.table}

\item{initial_days}{integer}
}
\value{
data.table
}
\description{
Create Initial ATC Indicators (New Enrollees)
}
