% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_end_log_file.R
\name{start_log_file}
\alias{start_log_file}
\title{Start a log file}
\usage{
start_log_file(log_file = TRUE, file_name = NULL, print = TRUE)
}
\arguments{
\item{log_file}{logical, if TRUE then starts a log file}

\item{file_name}{path to log file (relative to current directory)}
}
\description{
Start a log file
}
