% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_sample.R
\name{subset_sample}
\alias{subset_sample}
\title{Subset to analystic sample}
\usage{
subset_sample(DT, subset_vars, progress = F, balance_vars = NULL)
}
\arguments{
\item{DT}{a data.table}

\item{subset_vars}{variables to subset}

\item{progress}{logical, if TRUE prints where in subsetting process the
function is}

\item{balance_vars}{string character, if non-NULL, then estimates balance}
}
\value{
a list with three elements
}
\description{
Subset to analystic sample
}
