% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_opt.R
\name{unpack_opt}
\alias{unpack_opt}
\title{Unpack command line options}
\usage{
unpack_opt(option_list)
}
\arguments{
\item{option_list}{option list from command line}

\item{print}{logical, if TRUE then checks for common elemtns of option_list 
and prints them}
}
\description{
Unpack command line options
}
